+++ Libplanet.Net.Tests/Messages/MessageTest.cs
@@ -2,12 +2,15 @@ using System;
 using System.Collections.Immutable;
 using System.Linq;
 using System.Net;
+using Bencodex;
 using Libplanet.Blocks;
 using Libplanet.Consensus;
 using Libplanet.Crypto;
+using Libplanet.Net.Consensus;
 using Libplanet.Net.Messages;
 using Libplanet.Net.Transports;
 using Libplanet.Tests.Common.Action;
+using Libplanet.Tests.Store;
 using NetMQ;
 using Xunit;
 using static Libplanet.Tests.TestUtils;
@@ -149,6 +152,29 @@ namespace Libplanet.Net.Tests.Messages
             // Invalid message cases
             Assert.Throws<InvalidMessageException>(() => new ConsensusPreVoteMsg(preCommit));
             Assert.Throws<InvalidMessageException>(() => new ConsensusPreCommitMsg(preVote));
+        }
+
+        [Fact]
+        public void DifferentBlockHashProposal()
+        {
+            var codec = new Codec();
+            var fx = new MemoryStoreFixture();
+            var key = new PrivateKey();
+
+            var baseProposal = new ProposalMetaData(
+                1,
+                0,
+                codec.Encode(fx.Block1.MarshalBlock()),
+                fx.Block1.Timestamp,
+                key.PublicKey,
+                -1).Sign(key);
+
+            // Valid message case
+            _ = new ConsensusProposeMsg(key.PublicKey, 1, 0, fx.Block1.Hash, baseProposal);
+
+            // Invalid message case
+            Assert.Throws<ArgumentException>(() =>
+                new ConsensusProposeMsg(key.PublicKey, 1, 0, fx.Block2.Hash, baseProposal));
         }
     }
 }